refactor: remove file "types.ts"
- remove file "types.ts"
- move all types from "type.ts" into appropriate files

closes #406
