+++ packages/mongodb-memory-server-core/src/util/MongoBinaryDownload.ts
@@ -8,7 +8,6 @@ import { createUnzip } from 'zlib';
 import tar from 'tar-stream';
 import yauzl from 'yauzl';
 import MongoBinaryDownloadUrl from './MongoBinaryDownloadUrl';
-import { DownloadProgress } from '../types';
 import { LATEST_VERSION } from './MongoBinary';
 import { HttpsProxyAgent } from 'https-proxy-agent';
 import { promisify } from 'util';
@@ -16,6 +15,13 @@ import resolveConfig, { envToBool } from './resolve-config';
 import debug from 'debug';
 
 const log = debug('MongoMS:MongoBinaryDownload');
+
+export interface DownloadProgress {
+  current: number;
+  length: number;
+  totalMb: number;
+  lastPrintedAt: number;
+}
 
 export interface MongoBinaryDownloadOpts {
   version?: string;
